Add module for addRemove button for Project
