--- modules/removeBtn.js
@@ -1,4 +1,0 @@
-const removeBooks = (id,arr)=> {
-    let array = arr.filter((e) => e.id !== id);
-    return array
-}
